+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Drop support for Node 18 and below (they will likely still work, but are no longer
   tested)
+- Export `isDecoder(value)` helper
 
 ## [2.7.5] - 2025-09-02